+++ design-toolkit/components/src/components/classification-badge/types.ts
@@ -10,7 +10,7 @@
  * governing permissions and limitations under the License.
  */
 
-import type { HTMLAttributes, PropsWithChildren, RefAttributes } from 'react';
+import type { ComponentPropsWithRef } from 'react';
 import type { VariantProps } from 'tailwind-variants';
 import type { ClassificationBadgeStyles } from './styles';
 
@@ -19,8 +19,6 @@ export type ClassificationBadgeStyleVariants = VariantProps<
 >;
 
 export type ClassificationBadgeProps = ClassificationBadgeStyleVariants &
-  RefAttributes<HTMLSpanElement> &
-  PropsWithChildren<HTMLAttributes<HTMLSpanElement>>;
-
-export type ClassificationBadgeProviderProps =
-  PropsWithChildren<ClassificationBadgeProps>;
+  ComponentPropsWithRef<'span'> & {
+    size?: 'medium' | 'small';
+  };