+++ design-toolkit/components/src/components/color-picker/color-picker.test.tsx
@@ -12,9 +12,10 @@
 
 import { render, screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
-import { ColorPicker, type ColorPickerProps } from './index';
+import { ColorPicker } from './';
+import type { ColorPickerProps } from './types';
 
-const options = [
+const items = [
   '#ECECE6',
   '#898989',
   '#62a6ff',
@@ -23,10 +24,10 @@ const options = [
   '#D4231D',
 ];
 
-function setup({ ...rest }: Partial<ColorPickerProps> = {}) {
-  render(<ColorPicker options={options} {...rest} />);
+function setup(props: Partial<ColorPickerProps> = {}) {
   return {
-    ...rest,
+    ...render(<ColorPicker items={items} {...props} />),
+    ...props,
   };
 }