+++ design-toolkit/components/src/components/menu/index.tsx
@@ -18,7 +18,6 @@ import ChevronRight from '@accelint/icons/chevron-right';
 import { createContext, useContext } from 'react';
 import {
   Header as AriaHeader,
-  Keyboard as AriaKeyboard,
   Menu as AriaMenu,
   Collection as AriaMenuCollection,
   MenuItem as AriaMenuItem,
@@ -28,16 +27,17 @@ import {
   SubmenuTrigger as AriaSubmenuTrigger,
   Text as AriaText,
   type ContextValue,
+  DEFAULT_SLOT,
+  KeyboardContext,
   Popover,
+  Provider,
   composeRenderProps,
   useContextProps,
 } from 'react-aria-components';
-import { Icon } from '../icon';
+import { Icon, IconContext } from '../icon';
 import { MenuStyles, MenuStylesDefaults } from './styles';
 import type {
-  MenuIconProps,
   MenuItemProps,
-  MenuKeyboardProps,
   MenuProps,
   MenuSectionProps,
   MenuTextProps,
@@ -53,93 +53,14 @@ const {
   more,
   sectionHeader,
   separator,
-  keyboard,
+  hotkey,
+  popover,
 } = MenuStyles();
 
 export const MenuContext =
   createContext<ContextValue<MenuProps<unknown>, HTMLDivElement>>(null);
 
-export function Menu<T extends object>({ ref, ...props }: MenuProps<T>) {
-  [props, ref] = useContextProps(props, ref ?? null, MenuContext);
-
-  const {
-    children,
-    className,
-    selectionMode = 'single',
-    offset = -4,
-    placement,
-    isNonModal,
-    containerPadding,
-    variant = MenuStylesDefaults.variant,
-    ...rest
-  } = props;
-
-  return (
-    <Popover
-      offset={offset}
-      placement={placement}
-      isNonModal={isNonModal}
-      containerPadding={containerPadding}
-    >
-      <MenuContext.Provider value={{ variant }}>
-        <AriaMenu
-          ref={ref}
-          className={composeRenderProps(className, (className) =>
-            menu({ className, variant }),
-          )}
-          selectionMode={selectionMode}
-          {...rest}
-        >
-          {children}
-        </AriaMenu>
-      </MenuContext.Provider>
-    </Popover>
-  );
-}
-Menu.displayName = 'Menu';
-
-export const MenuItem = (props: MenuItemProps) => {
-  const context = useContext(MenuContext);
-  const variant =
-    (isSlottedContextValue(context) ? undefined : context?.variant) ??
-    MenuStylesDefaults.variant;
-
-  const {
-    children: childrenProp,
-    classNames,
-    color = MenuStylesDefaults.color,
-    ...rest
-  } = props;
-
-  return (
-    <AriaMenuItem
-      {...rest}
-      className={composeRenderProps(classNames?.item, (className) =>
-        item({ className, variant, color }),
-      )}
-    >
-      {composeRenderProps(props.children, (children, { hasSubmenu }) => (
-        <>
-          {typeof children === 'string' ? (
-            <AriaText className={classNames?.text} slot='label'>
-              {children}
-            </AriaText>
-          ) : (
-            children
-          )}
-          {hasSubmenu && (
-            <Icon className={more({ className: classNames?.more })}>
-              <ChevronRight />
-            </Icon>
-          )}
-        </>
-      ))}
-    </AriaMenuItem>
-  );
-};
-MenuItem.displayName = 'Menu.Item';
-
-export function MenuSection<T extends object>(props: MenuSectionProps<T>) {
+function MenuSection<T extends object>(props: MenuSectionProps<T>) {
   const { header, children, classNames, items, ...rest } = props;
 
   return (
@@ -155,12 +76,12 @@ export function MenuSection<T extends object>(props: MenuSectionProps<T>) {
 }
 MenuSection.displayName = 'Menu.Section';
 
-export function MenuSeparator({ className, ...rest }: SeparatorProps) {
+function MenuSeparator({ className, ...rest }: SeparatorProps) {
   return <AriaSeparator {...rest} className={separator({ className })} />;
 }
 MenuSeparator.displayName = 'Menu.Separator';
 
-export function MenuLabel(props: MenuTextProps) {
+function MenuLabel(props: MenuTextProps) {
   const { children, className, ...rest } = props;
 
   return (
@@ -171,14 +92,13 @@ export function MenuLabel(props: MenuTextProps) {
 }
 MenuLabel.displayName = 'Menu.Item.Label';
 
-export function MenuDescription(props: MenuTextProps) {
+function MenuDescription(props: MenuTextProps) {
   const { children, className, ...rest } = props;
 
   return (
     <AriaText
       {...rest}
       slot='description'
-      data-slot='description'
       className={description({ className })}
     >
       {children}
@@ -187,34 +107,104 @@ export function MenuDescription(props: MenuTextProps) {
 }
 MenuDescription.displayName = 'Menu.Item.Description';
 
-export function MenuItemIcon(props: MenuIconProps) {
-  const { children, className, ...rest } = props;
+function MenuItem(props: MenuItemProps) {
+  const context = useContext(MenuContext);
+  const variant =
+    (isSlottedContextValue(context) ? undefined : context?.variant) ??
+    MenuStylesDefaults.variant;
+
+  const {
+    classNames,
+    color = MenuStylesDefaults.color,
+    children,
+    ...rest
+  } = props;
 
   return (
-    <Icon {...rest} className={icon({ className })}>
-      {children}
-    </Icon>
+    <AriaMenuItem
+      {...rest}
+      className={composeRenderProps(classNames?.item, (className) =>
+        item({ className, variant, color }),
+      )}
+    >
+      {composeRenderProps(children, (children, { hasSubmenu }) => (
+        <Provider
+          values={[
+            [
+              KeyboardContext,
+              { className: hotkey({ className: classNames?.hotkey }) },
+            ],
+            [
+              IconContext,
+              {
+                slots: {
+                  [DEFAULT_SLOT]: {
+                    className: icon({ className: classNames?.icon }),
+                  },
+                  submenu: { className: more({ className: classNames?.more }) },
+                },
+              },
+            ],
+          ]}
+        >
+          {typeof children === 'string' ? (
+            <AriaText className={classNames?.text} slot='label'>
+              {children}
+            </AriaText>
+          ) : (
+            children
+          )}
+          {hasSubmenu && (
+            <Icon slot='submenu'>
+              <ChevronRight />
+            </Icon>
+          )}
+        </Provider>
+      ))}
+    </AriaMenuItem>
   );
 }
-MenuItemIcon.displayName = 'Menu.Item.Icon';
+MenuItem.displayName = 'Menu.Item';
+MenuItem.Label = MenuLabel;
+MenuItem.Description = MenuDescription;
 
-export function MenuItemKeyboard(props: MenuKeyboardProps) {
-  const { children, className, ...rest } = props;
+export function Menu<T extends object>({ ref, ...props }: MenuProps<T>) {
+  [props, ref] = useContextProps(props, ref ?? null, MenuContext);
+
+  const {
+    children,
+    classNames,
+    selectionMode = 'single',
+    variant = MenuStylesDefaults.variant,
+    popoverProps,
+    ...rest
+  } = props;
 
   return (
-    <AriaKeyboard {...rest} className={keyboard({ className })}>
-      {children}
-    </AriaKeyboard>
+    <Popover
+      {...popoverProps}
+      className={composeRenderProps(classNames?.popover, (className) =>
+        popover({ className }),
+      )}
+    >
+      <MenuContext.Provider value={{ variant }}>
+        <AriaMenu
+          ref={ref}
+          className={composeRenderProps(classNames?.menu, (className) =>
+            menu({ className, variant }),
+          )}
+          selectionMode={selectionMode}
+          {...rest}
+        >
+          {children}
+        </AriaMenu>
+      </MenuContext.Provider>
+    </Popover>
   );
 }
-MenuItemKeyboard.displayName = 'Menu.Item.Keyboard';
-
+Menu.displayName = 'Menu';
 Menu.Trigger = AriaMenuTrigger;
 Menu.Submenu = AriaSubmenuTrigger;
 Menu.Item = MenuItem;
-MenuItem.Icon = MenuItemIcon;
-MenuItem.Label = MenuLabel;
-MenuItem.Description = MenuDescription;
-MenuItem.Keyboard = MenuItemKeyboard;
 Menu.Separator = MenuSeparator;
 Menu.Section = MenuSection;