+++ design-toolkit/components/src/components/query-builder/value-selector.tsx
@@ -10,16 +10,18 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
+'use client';
 
+import 'client-only';
 import type { Key } from '@react-types/shared';
 import { useCallback, useMemo } from 'react';
-import { Header, ListBoxSection } from 'react-aria-components';
 import {
   type ValueSelectorProps,
   isOptionGroupArray,
   useValueSelector,
 } from 'react-querybuilder';
-import { ComboBox } from '../combobox';
+import { ComboBoxField } from '../combobox-field';
+import { Options } from '../options';
 
 export function ValueSelector(props: ValueSelectorProps) {
   const {
@@ -54,34 +56,28 @@ export function ValueSelector(props: ValueSelectorProps) {
     () =>
       isOptionGroupArray(optionsProp)
         ? optionsProp.map((section) => (
-            <ListBoxSection key={section.label}>
-              <Header>{section.label}</Header>
+            <Options.Section key={section.label} header={section.label}>
               {section.options.map((option) => (
-                <ComboBox.Item
-                  name={option.name}
-                  id={option.name}
-                  key={option.name}
-                >
+                <Options.Item id={option.name} key={option.name}>
                   {option.label}
-                </ComboBox.Item>
+                </Options.Item>
               ))}
-            </ListBoxSection>
+            </Options.Section>
           ))
         : optionsProp.map((option) => (
-            <ComboBox.Item
-              name={option.name}
+            <Options.Item
               textValue={option.label}
               id={option.name}
               key={option.name}
             >
               {option.label}
-            </ComboBox.Item>
+            </Options.Item>
           )),
     [optionsProp],
   );
 
   return (
-    <ComboBox
+    <ComboBoxField
       size='small'
       isDisabled={disabled}
       {...rest}
@@ -90,6 +86,6 @@ export function ValueSelector(props: ValueSelectorProps) {
       onSelectionChange={handleSelectionChange}
     >
       {options}
-    </ComboBox>
+    </ComboBoxField>
   );
 }