+++ design-toolkit/components/src/components/text-area-field/index.tsx
@@ -13,6 +13,7 @@
 'use client';
 
 import 'client-only';
+import type { ProviderProps } from '@/lib/types';
 import { createContext } from 'react';
 import {
   type ContextValue,
@@ -24,18 +25,18 @@ import {
   useContextProps,
 } from 'react-aria-components';
 import { Label } from '../label';
-import { TextAreaStyles, TextAreaStylesDefaults } from './styles';
-import type { TextAreaFieldProps, TextAreaFieldProviderProps } from './types';
+import { TextAreaStyles } from './styles';
+import type { TextAreaFieldProps } from './types';
 
 const { field, label, input, description, error } = TextAreaStyles();
 
 export const TextAreaFieldContext =
-  createContext<ContextValue<TextAreaFieldProviderProps, HTMLDivElement>>(null);
+  createContext<ContextValue<TextAreaFieldProps, HTMLDivElement>>(null);
 
 function TextAreaFieldProvider({
   children,
   ...props
-}: TextAreaFieldProviderProps) {
+}: ProviderProps<TextAreaFieldProps>) {
   return (
     <TextAreaFieldContext.Provider value={props}>
       {children}
@@ -52,45 +53,31 @@ export function TextAreaField({ ref, ...props }: TextAreaFieldProps) {
     description: descriptionProp,
     errorMessage: errorMessageProp,
     label: labelProp,
-    inputRef,
     inputProps,
-    size = TextAreaStylesDefaults.size,
-    isDisabled,
+    size = 'medium',
     isInvalid: isInvalidProp,
-    isReadOnly,
-    isRequired,
     ...rest
   } = props;
   const errorMessage = errorMessageProp || null; // Protect against empty string
-  const isInvalid = isInvalidProp ?? (errorMessage ? true : undefined); // Leave uncontrolled if possible to fallback to validation state
   const isSmall = size === 'small';
 
   return (
     <TextField
       {...rest}
-      className={composeRenderProps(
-        classNames?.field,
-        (className, { isInvalid }) =>
-          field({ className, size, isDisabled, isInvalid, isReadOnly }),
+      ref={ref}
+      className={composeRenderProps(classNames?.field, (className) =>
+        field({ className }),
       )}
-      isDisabled={isDisabled}
-      isInvalid={isInvalid}
-      isReadOnly={isReadOnly}
-      isRequired={isRequired}
+      isInvalid={isInvalidProp || (errorMessage ? true : undefined)} // Leave uncontrolled if possible to fallback to validation state
+      data-size={size}
     >
       {(
-        { isInvalid }, // Rely on internal state, not props, since validation result could differ from props
+        { isDisabled, isInvalid, isRequired }, // Rely on internal state, not props, since state could differ from props
       ) => (
         <>
-          {!isSmall && !!labelProp && (
+          {!!labelProp && !isSmall && (
             <Label
-              className={label({
-                className: classNames?.label,
-                size,
-                isDisabled,
-                isInvalid,
-                isReadOnly,
-              })}
+              className={label({ className: classNames?.label })}
               isDisabled={isDisabled}
               isRequired={isRequired}
             >
@@ -98,41 +85,22 @@ export function TextAreaField({ ref, ...props }: TextAreaFieldProps) {
             </Label>
           )}
           <TextArea
-            ref={inputRef}
             {...inputProps}
             className={composeRenderProps(classNames?.input, (className) =>
-              input({
-                className,
-                size,
-                isDisabled,
-                isInvalid,
-                isReadOnly,
-              }),
+              input({ className }),
             )}
           />
-          {!(isSmall || isInvalid) && !!descriptionProp && (
+          {!!descriptionProp && !(isSmall || isInvalid) && (
             <Text
               slot='description'
-              className={description({
-                className: classNames?.description,
-                size,
-                isDisabled,
-                isInvalid,
-                isReadOnly,
-              })}
+              className={description({ className: classNames?.description })}
             >
               {descriptionProp}
             </Text>
           )}
           <FieldError
             className={composeRenderProps(classNames?.error, (className) =>
-              error({
-                className,
-                size,
-                isDisabled,
-                isInvalid,
-                isReadOnly,
-              }),
+              error({ className }),
             )}
           >
             {errorMessage}