+++ apps/next/CHANGELOG.md
@@ -1,5 +1,72 @@
 # @apps/next
 
+## 0.2.26
+
+### Patch Changes
+
+- Updated dependencies [a97d386]
+- Updated dependencies [9ec4cbe]
+- Updated dependencies [83104ea]
+  - @accelint/design-toolkit@2.6.0
+  - @accelint/design-system@0.11.3
+  - @accelint/icons@2.0.3
+  - @accelint/bus@0.1.2
+
+## 0.2.25
+
+### Patch Changes
+
+- Updated dependencies [11a813f]
+- Updated dependencies [5404e48]
+- Updated dependencies [ca3922a]
+- Updated dependencies [7481c80]
+- Updated dependencies [a201c50]
+- Updated dependencies [de8c60a]
+- Updated dependencies [61a55ef]
+- Updated dependencies [484c14f]
+- Updated dependencies [a16fbf3]
+- Updated dependencies [c95e5b4]
+  - @accelint/design-toolkit@2.5.0
+  - @accelint/icons@2.0.2
+  - @accelint/bus@0.1.1
+  - @accelint/design-system@0.11.2
+
+## 0.2.24
+
+### Patch Changes
+
+- Updated dependencies [382bbc2]
+  - @accelint/design-toolkit@2.4.3
+
+## 0.2.23
+
+### Patch Changes
+
+- Updated dependencies [93b3750]
+  - @accelint/design-toolkit@2.4.2
+
+## 0.2.22
+
+### Patch Changes
+
+- Updated dependencies [f57cb86]
+- Updated dependencies [4fa066a]
+  - @accelint/design-toolkit@2.4.1
+
+## 0.2.21
+
+### Patch Changes
+
+- Updated dependencies [4dd70f2]
+- Updated dependencies [fe10a47]
+- Updated dependencies [c16b479]
+- Updated dependencies [51422c1]
+- Updated dependencies [a502977]
+- Updated dependencies [9421185]
+- Updated dependencies [34f3d91]
+- Updated dependencies [f9e0cac]
+  - @accelint/design-toolkit@2.4.0
+
 ## 0.2.20
 
 ### Patch Changes