+++ packages/bus/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @accelint/bus
 
+## 0.1.2
+
+### Patch Changes
+
+- 83104ea: Refactored ViewStack to be event driven, allowing for triggers anywhere in the app
+
+## 0.1.1
+
+### Patch Changes
+
+- ca3922a: added subpath exports for packages
+
 ## 0.1.0
 
 ### Minor Changes